+++ example/lib/shared/player/knight.dart
@@ -97,6 +97,7 @@ class Knight extends SimplePlayer
 
   void execRangeAttack(double angle, double damage) {
     this.simpleAttackRangeByAngle(
+      attackFrom: AttackFromEnum.PLAYER_OR_ALLY,
       animation: CommonSpriteSheet.fireBallRight,
       animationDestroy: CommonSpriteSheet.explosionAnimation,
       angle: angle,
@@ -133,11 +134,12 @@ class Knight extends SimplePlayer
   }
 
   @override
-  void receiveDamage(double damage, dynamic from) {
-    super.receiveDamage(damage, from);
-    if (hasController) {
+  bool checkCanReceiveDamage(AttackFromEnum attacker, double damage, from) {
+    bool shouldReceive = super.checkCanReceiveDamage(attacker, damage, from);
+    if (shouldReceive && hasController) {
       controller.onReceiveDamage(damage);
     }
+    return shouldReceive;
   }
 
   void execShowEmote() {