+++ lib/util/extensions/extensions.dart
@@ -7,6 +7,7 @@ import 'package:bonfire/camera/bonfire_camera.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/widgets.dart' as widget;
 
+export 'ally/ally_extensions.dart';
 export 'attackable_extensions.dart';
 export 'enemy/enemy_extensions.dart';
 export 'enemy/enemy_extensions.dart';
@@ -14,6 +15,7 @@ export 'enemy/rotation_enemy_extensions.dart';
 export 'game_component_extensions.dart';
 export 'joystick_extensions.dart';
 export 'movement_extensions.dart';
+export 'npc/npc_extensions.dart';
 export 'player/player_extensions.dart';
 export 'player/rotation_player_extensions.dart';
 
@@ -153,7 +155,7 @@ extension SpriteExt on Sprite {
 extension SpriteFutureExt on Future<Sprite> {
   Future<SpriteAnimation> toAnimation() async {
     var sprite = await this;
-    return SpriteAnimation.spriteList([sprite], stepTime: 0);
+    return SpriteAnimation.spriteList([sprite], stepTime: 1);
   }
 }